refactor(event-runtime): split API route modules (WM-290)
